requirements.txt
accelerate==0.25.0
certifi==2023.11.17
charset-normalizer==3.3.2
cmake==3.28.1
compel==2.0.2
diffusers==0.24.0
filelock==3.13.1
fsspec==2023.12.2
huggingface-hub==0.20.1
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.2
lit==17.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
psutil==5.9.7
pyparsing==3.1.1
pyre-extensions==0.0.29
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
sympy==1.12
tokenizers==0.15.0
torch==2.0.1
torchdeltacnn==0.0.0
torchvision==0.15.2
tqdm==4.66.1
transformers==4.36.2
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
xformers==0.0.20
zipp==3.17.0
